setup youcompleteme Respect local/global installation options.
Sort out indentation.

+++ dotfiles/vimrc/setup_youcompleteme.sh
@@ -49,6 +49,10 @@ case "${uname_result}" in
     echo "Platform result suggests Mingw64 without Msys2/Mingw64 terminal as required. Please install Msys2 and launch a Mingw64 console from there."
     echo "Platform result: '${uname_result}'"
     ;;
+  *linuxkit*)
+    platform=linux_kit
+    sub_platform=linux_kit
+    ;;
   *)
     echo 'Unrecognised platform ' >&2
     uname -a
@@ -57,7 +61,11 @@ esac
 
 function install_llvm() {
   local do_llvm
+  local do_global_installs
+  local do_user_installs
   do_llvm=${1:-1}
+  do_global_installs=${2:-1}
+  do_user_installs=${3:-1}
 
   case "${platform}" in
     darwin)
@@ -110,8 +118,13 @@ function install_llvm() {
 
 function install_python() {
   if [ ${1:-1} -ne 1 ]; then return; fi
+  local do_global_installs
+  local do_user_installs
 
-  if ! which python3 2>&1 1>/dev/null
+  do_global_installs=${2:-1}
+  do_user_installs=${3:-1}
+
+  if ! command -v python3 2>&1 1>/dev/null
   then
     echo 'Installing python3'
     case "${platform}" in
@@ -130,11 +143,14 @@ function install_python() {
         esac
         ;;
       *)
-        echo "Don't know how to install python3 on this platform: $(uname -a)"
+        if [ ${do_global_installs} -ne 0 ]
+        then
+          echo "Don't know how to install python3 on this platform: $(uname -a)"
+        fi
         ;;
     esac
 
-    if ! which python3 2>&1 1>/dev/null
+    if [ ${do_global_installs} -ne 0 ] && ! command -v python3 2>&1 1>/dev/null
     then
       echo 'Failed to install python3 !'
       exit 1
@@ -220,6 +236,12 @@ function detect_python() {
 }
 
 function install_cmake() {
+  local do_global_installs
+  local do_user_installs
+
+  do_global_installs=${1:-1}
+  do_user_installs=${2:-1}
+
   # For compilers to find llvm you may need to set:
   #   export LDFLAGS="-L/usr/local/opt/llvm/lib"
   #   export CPPFLAGS="-I/usr/local/opt/llvm/include"
@@ -235,6 +257,16 @@ function install_cmake() {
         #pacman -Ss cmake
         pacman -Ss mingw64/mingw-w64-x86_64-cmake
         ;;
+      linux_kit)
+        if [ ${do_global_installs} -ne 0 ]
+        then
+          if command -v microdnf >/dev/null
+          then
+            microdnf --assumeyes update && microdnf --assumeyes install cmake
+          else
+            dnf --assumeyes update && microdnf --assumeyes install cmake
+          fi
+        fi # do_global_installs
     esac
   fi
 }
@@ -244,6 +276,14 @@ YCM_THIRDPARTY_DIR="${YCM_DIR}/third_party/ycmd"
 
 function mend_ycm() {
   if [ ${1:-1} -ne 1 ]; then return; fi
+  local do_user_installs
+  do_user_installs=${2:-1}
+
+  if [ ${do_user_installs} -eq 0 ]
+  then
+    echo "Skipping mending YouCompleteMe code" >&2
+    return 0
+  fi
 
   if [ -d "${YCM_THIRDPARTY_DIR}" ]
   then
@@ -259,6 +299,15 @@ function mend_ycm() {
 }
 
 function fixup_cmake_files() {
+  local do_user_installs
+  do_user_installs=${1:-1}
+
+  if [ ${do_user_installs} -eq 0 ]
+  then
+    echo "Skipping fixing YouCompleteMe cmake files" >&2
+    return 0
+  fi
+
   case "${platform}" in
     darwin)
       if ! grep 'CMAKE_OSX_DEPLOYMENT_TARGET' "${YCM_THIRDPARTY_DIR}/cpp/CMakeLists.txt"  >/dev/null
@@ -299,6 +348,15 @@ EOF
 
 function build_ycm_core() {
   if [ ${1:-1} -ne 1 ]; then return; fi
+  local do_user_installs
+
+  do_user_installs=${2:-1}
+
+  if [ ${do_user_installs} -eq 0 ]
+  then
+    echo "Skipping building YouCompleteMe core" >&2
+    return 0
+  fi
 
   local system_clang_path
   local build_dir
@@ -349,6 +407,14 @@ function build_ycm_core() {
 
 function build_ycm_regex() {
   if [ ${1:-1} -ne 1 ]; then return; fi
+  local do_user_installs
+  do_user_installs=${2:-1}
+
+  if [ ${do_user_installs} -eq 0 ]
+  then
+    echo "Skipping building YouCompleteMe regex." >&2
+    return 0
+  fi
 
   echo 'Setting up regex build directories.'
   rm -rf ~/stuff/builds/ycm/regex_build/${sub_platform}
@@ -365,7 +431,13 @@ function build_ycm_regex() {
 
 # node.js
 function install_node() {
-  if ! which npm 2>&1 1>/dev/null
+  local do_global_installs
+  local do_user_installs
+
+  do_global_installs=${1:-1}
+  do_user_installs=${2:-1}
+
+  if ! command -v npm 2>&1 1>/dev/null
   then
     echo 'Installing npm...'
     case "${platform}" in
@@ -374,7 +446,13 @@ function install_node() {
         ;;
       msys)
         mkdir -p "{$HOME}/code/thirdparty/node.js"
-        (cd "${HOME}/code/thirdparty/node.js"; git clone git@github.com:nodejs/node.git node.git; cd node.git; echo 'Need to follow instructions in BUILDING.md')
+        (cd "${HOME}/code/thirdparty/node.js"; git clone "$(_github_repo_address nodejs/node.git)" node.git; cd node.git; echo 'Need to follow instructions in BUILDING.md')
+        ;;
+      *)
+        if [ ${do_global_installs} -ne 0 ]
+        then
+          echo "Skipping node installation on ${platform}" >&2
+        fi
         ;;
     esac
   fi
@@ -441,8 +519,13 @@ function verify_java() {
 
 function install_rust() {
   if [ ${1:-1} -ne 1 ]; then return; fi
+  local do_global_installs
+  local do_user_installs
 
-  if ! which cargo &>/dev/null
+  do_global_installs=${2:-1}
+  do_user_installs=${3:-1}
+
+  if ! command -v cargo &>/dev/null
   then
     echo 'Installing rust...'
     case "${platform}" in
@@ -457,12 +540,15 @@ function install_rust() {
         #pacman -Su mingw64/mingw-w64-x86_64-rust
         ;;
       *)
-        echo "Don't know how to install rust on platform '${platform}'"
+        if [ ${do_global_installs} -ne 0 ]
+        then
+          echo "Don't know how to install rust on platform '${platform}'"
+        fi
         ;;
     esac
   fi
 
-  if ! which cargo &>/dev/null
+  if ! command -v cargo &>/dev/null
   then
     echo 'Failed to install rust.' >&2
   fi
@@ -472,13 +558,21 @@ function build_extra_components() {
   local do_java
   local do_node
   local do_rust
+  local do_user_installs
   local install_args
 
   do_java=${1:-1}
   do_node=${2:-1}
   do_rust=${3:-1}
+  do_user_installs=${4:-1}
   declare -a install_args
   install_args=()
+
+  if [ ${do_user_installs} -eq 0 ]
+  then
+    echo "Skipping building extra YouCompleteMe components." >&2
+    return 0
+  fi
 
   if [ ${do_java} -ne 0 ]
   then
@@ -524,34 +618,79 @@ function build_extra_components() {
 #  esac
 }
 
-function install_bear() {
-  if ! which bear 2>&1 1>/dev/null
+function _github_repo_address() {
+  local input
+  input="${1}"
+
+  if [ -z "${input}" ]
   then
-    echo 'Installing bear'
-    rm -rf ~/code/thirdparty/bear/build
-    mkdir -p ~/code/thirdparty/bear/build
-    pushd ~/code/thirdparty/bear
-    git clone git@github.com:rizsotto/Bear.git
-    pushd ~/code/thirdparty/bear/build
-    cmake ../Bear
-    make all
-    make insatll
-    echo 'Installed bear.'
-    which bear
-    popd
+    echo 'Expected input for _github_repo_address' >&2
+    return 1
+  elif [ -d "${HOME}/.ssh" ] && find "${HOME}/.ssh" -name '*.pub' | grep --quiet '.'
+  then
+    echo "git@github.com:${input}"
+  else
+    echo "https://github.com/${input}"
   fi
 }
 
-function install_intercept_build() {
-  if ! which intercept-build 2>&1 1>/dev/null
+function install_bear() {
+  if [ ${1:-1} -ne 1 ]; then return 0; fi
+
+  local do_global_installs
+  local do_user_installs
+  do_global_installs=${2:-1}
+  do_user_installs=${3:-1}
+
+  if [ ${do_global_installs} -ne 0 ]
   then
-    echo 'Installing intercept-build'
-    pip3 install --upgrade scan-build
-    if ! which intercept-build 2>&1 1>/dev/null
+    if ! command -v bear 2>&1 1>/dev/null
     then
-      echo 'Warning - failed to install intercept-build using "pip3 install scan-build"'
+      if [ ! -d "${HOME}/code/thirdparty/bear/Bear.git" ]
+      then
+        echo 'Downloading bear'
+        mkdir -p "${HOME}/code/thirdparty/bear"
+        pushd "${HOME}/code/thirdparty/bear" >/dev/null
+        git clone "$(_github_repo_address rizsotto/Bear.git)"
+      else
+        echo 'Updating bear'
+        pushd "${HOME}/code/thirdparty/bear/Bear.git" >/dev/null
+        git fetch --prune
+        git rebase
+      fi
+      popd >/dev/null
+      rm -rf "${HOME}/code/thirdparty/bear/build"
+      mkdir -p "${HOME}/code/thirdparty/bear/build"
+      pushd "${HOME}/code/thirdparty/bear/build" >/dev/null
+      echo 'Installing bear'
+      cmake ../Bear
+      make all
+      make install
+      echo 'Installed bear.'
+      command -v bear
+      popd >/dev/null
     fi
-  fi
+  fi # do_global_installs
+}
+
+function install_intercept_build() {
+  local do_global_installs
+  local do_user_installs
+  do_global_installs=${1:-1}
+  do_user_installs=${2:-1}
+
+  if [ ${do_global_installs} -ne 0 ]
+  then
+    if ! command -v intercept-build 2>&1 1>/dev/null
+    then
+      echo 'Installing intercept-build'
+      pip3 install --upgrade scan-build
+      if ! command -v intercept-build 2>&1 1>/dev/null
+      then
+        echo 'Warning - failed to install intercept-build using "pip3 install scan-build"'
+      fi
+    fi
+  fi # do_global_installs
 }
 
 do_python=1
@@ -563,6 +702,9 @@ do_mend_ycm=1
 do_ycm_regex=1
 do_override_mingw_python3_on_msys=1
 do_llvm=1
+do_bear=1
+do_global_installs=1
+do_user_installs=1
 
 case "${platform}" in
   msys)
@@ -577,6 +719,17 @@ case "${platform}" in
         do_mend_ycm=0
         ;;
     esac
+    ;;
+  linux_kit)
+    do_bear=0
+    if [ $(whoami) = 'root' ]
+    then
+      do_global_installs=1
+      do_user_installs=0
+    else
+      do_global_installs=0
+      do_user_installs=1
+    fi
     ;;
   *)
     :
@@ -606,24 +759,30 @@ function usage() {
 $(basename "${0}") [--[no-]python] [--[no-]java] [--[no-]node] [--[no-]rust] [--no-ycm] [--no-ycm-regex]
 
 
-  --python        If specified, install Python3 if necessary. Defaults to $(boolstring ${do_python}).
-  --java          If specified, install the java support in YouCompleteMe. Defaults to $(boolstring ${do_java}).
-  --node          If specified, install Node, and the plugin for YouCompleteMe. Defaults to $(boolstring ${do_node}).
-  --rust          If specified, install rust, and the plugin for YouCompleteMe. Defaults to $(boolstring ${do_rust}).
-  --mend-ycm      If specified, run 'git submodule update --init --recursive' after installing YouCompleteMe. Defaults to $(boolstring ${do_mend_ycm}).
-  --override-mingw-with-msys
-                  If specified on mingw platform with msys python3 available, build against msys platform. Defaults to $(boolstring ${do_override_mingw_python3_on_msys}).
+  --python             If specified, install Python3 if necessary. Defaults to $(boolstring ${do_python}).
+  --java               If specified, install the java support in YouCompleteMe. Defaults to $(boolstring ${do_java}).
+  --node               If specified, install Node, and the plugin for YouCompleteMe. Defaults to $(boolstring ${do_node}).
+  --rust               If specified, install rust, and the plugin for YouCompleteMe. Defaults to $(boolstring ${do_rust}).
+  --mend-ycm           If specified, run 'git submodule update --init --recursive' after installing YouCompleteMe. Defaults to $(boolstring ${do_mend_ycm}).
+  --override-mingw     -with-msys
+                       If specified on mingw platform with msys python3 available, build against msys platform. Defaults to $(boolstring ${do_override_mingw_python3_on_msys}).
+  --bear               If specified, install bear. Defaults to $(boolstring ${do_bear}).
 
-  --no-python     Don't try to install Python3. Defaults to $(inverseboolstring ${do_python}).
-  --no-java       Don't install the java support in YouCompleteMe. Defaults to $(inverseboolstring ${do_java}).
-  --no-node       Don't install Node, and the plugin for YouCompleteMe. Defaults to $(inverseboolstring ${do_node}).
-  --no-rust       Don't install rust, and the plugin for YouCompleteMe. Defaults to $(inverseboolstring ${do_rust}).
-  --no mend-ycm   Don't run 'git submodule update --init --recursive' after installing YouCompleteMe. Defaults to $(inverseboolstring ${do_mend_ycm}).
-  --no-ycm        Don't build YouCompleteMe. Defaults to $(inverseboolstring ${do_ycm}).
-  --no-ycm-regex  Don't build YouCompleteMe regex. Defaults to $(inverseboolstring ${do_ycm_regex}).
+  --no-python          Don't try to install Python3. Defaults to $(inverseboolstring ${do_python}).
+  --no-java            Don't install the java support in YouCompleteMe. Defaults to $(inverseboolstring ${do_java}).
+  --no-node            Don't install Node, and the plugin for YouCompleteMe. Defaults to $(inverseboolstring ${do_node}).
+  --no-rust            Don't install rust, and the plugin for YouCompleteMe. Defaults to $(inverseboolstring ${do_rust}).
+  --no mend-ycm        Don't run 'git submodule update --init --recursive' after installing YouCompleteMe. Defaults to $(inverseboolstring ${do_mend_ycm}).
+  --no-ycm             Don't build YouCompleteMe. Defaults to $(inverseboolstring ${do_ycm}).
+  --no-ycm-regex       Don't build YouCompleteMe regex. Defaults to $(inverseboolstring ${do_ycm_regex}).
   --no-override-mingw-with-msys
-                  Don't override mingw python3 installation with msys. Defaults to $(inverseboolstring ${do_override_mingw_python3_on_msys}).
-  --no-llvm       Don't install llvm. Defaults to $(inverseboolstring ${do_llvm}).
+                       Don't override mingw python3 installation with msys. Defaults to $(inverseboolstring ${do_override_mingw_python3_on_msys}).
+  --no-bear            Don't install bear. Defaults to $(inverseboolstring ${do_bear}).
+  --no-llvm            Don't install llvm. Defaults to $(inverseboolstring ${do_llvm}).
+  --global-installs    If specified, install global dependencies.
+  --no-global-installs If specified, do not install global dependencies.
+  --user-installs      If specified, install user dependencies.
+  --no-user-installs   If specified, do not install user dependencies.
 
 Notes:
 Java, node and rust all work on Mac, but are in progress for Msys/Mingw64.
@@ -654,60 +813,79 @@ do
 
   case "${arg}" in
     --llvm)
-        do_llvm=1
-        ;;
+      do_llvm=1
+      ;;
     --python)
-        do_python=1
-        ;;
+      do_python=1
+      ;;
     --java)
-        do_java=1
-        ;;
+      do_java=1
+      ;;
     --node)
-        do_node=1
-        ;;
+      do_node=1
+      ;;
     --rust)
-        do_rust=1
-        ;;
+      do_rust=1
+      ;;
     --mend-ycm)
-        do_mend_ycm=1
-        ;;
+      do_mend_ycm=1
+      ;;
     --override-mingw-with-msys)
-        do_override_mingw_python3_on_msys=1
-        ;;
+      do_override_mingw_python3_on_msys=1
+      ;;
+    --bear)
+      do_bear=1
+      ;;
 
     --no-python)
-        do_python=0
-        ;;
+      do_python=0
+      ;;
     --no-java)
-        do_java=0
-        ;;
+      do_java=0
+      ;;
     --no-node)
-        do_node=0
-        ;;
+      do_node=0
+      ;;
     --no-rust)
-        do_rust=0
-        ;;
+      do_rust=0
+      ;;
     --no-mend-ycm)
-        do_mend_ycm=0
-        ;;
+      do_mend_ycm=0
+      ;;
     --no-ycm)
-        do_ycm=0
-        ;;
+      do_ycm=0
+      ;;
     --no-ycm-regex)
-        do_ycm_regex=0
-        ;;
+      do_ycm_regex=0
+      ;;
     --no-override-mingw-with-msys)
-        do_override_mingw_python3_on_msys=0
-        ;;
+      do_override_mingw_python3_on_msys=0
+      ;;
+    --no-bear)
+      do_bear=0
+      ;;
     --no-llvm)
-        do_llvm=0
-        ;;
+      do_llvm=0
+      ;;
+
+    --global-installs)
+      do_global_installs=1
+      ;;
+    --no-global-installs)
+      do_global_installs=0
+      ;;
+    --user-installs)
+      do_user_installs=1
+      ;;
+    --no-user-installs)
+      do_user_installs=0
+      ;;
     --help|-h|/?)
-        usage
-        exit 1
-        ;;
+      usage
+      exit 1
+      ;;
     *)
-      echo "Unexpected argument '${arg}'" >&2
+      echo "Ignoring argument '${arg}'" >&2
       (( ok_to_continue++ ))
       ;;
   esac
@@ -718,35 +896,49 @@ then
   exit ${ok_to_continue}
 fi
 
+if [ ${do_global_installs} -eq 0 ]
+then
+  global_installs_arg='--no-global-installs'
+else
+  global_installs_arg='--global-installs'
+fi
+
+if [ ${do_user_installs} -eq 0 ]
+then
+  user_installs_arg='--no-user-installs'
+else
+  user_installs_arg='--user-installs'
+fi
+
 case "${platform}" in
   msys)
     :
     ;;
   *)
-    install_llvm ${do_llvm}
+    install_llvm ${do_llvm} ${do_global_installs} ${do_user_installs}
     ;;
 esac
 
-install_python ${do_python}
-install_cmake
+install_python ${do_python} ${do_global_installs} ${do_user_installs}
+install_cmake ${do_global_installs} ${do_user_installs}
 detect_python ${do_override_mingw_python3_on_msys} 'building YouCompleteMe core'
-mend_ycm ${do_mend_ycm}
-fixup_cmake_files
-build_ycm_core ${do_ycm}
+mend_ycm ${do_mend_ycm} ${do_user_installs}
+fixup_cmake_files ${do_user_installs}
+build_ycm_core ${do_ycm} ${do_user_installs}
 detect_python 0 'building YouCompleteMe regex'
-build_ycm_regex ${do_ycm_regex}
-install_node ${do_node}
+build_ycm_regex ${do_ycm_regex} ${do_user_installs}
+install_node ${do_node} ${do_global_installs} ${do_user_installs}
 verify_java ${do_java}
-install_rust ${do_rust}
-build_extra_components ${do_java} ${do_node} ${do_rust}
+install_rust ${do_rust} ${do_global_installs} ${do_user_installs}
+build_extra_components ${do_java} ${do_node} ${do_rust} ${do_user_installs}
 case "${platform}" in
   msys)
     :
     ;;
   *)
-    install_bear
+    install_bear ${do_bear} ${do_global_installs} ${do_user_installs}
     ;;
 esac
-install_intercept_build
+install_intercept_build ${do_global_installs} ${do_user_installs}
 echo 'Done'
